taiwa_bot/requirements.txt
discord.py
openai
gspread
oauth2client
python-dotenv
pandas